tests/wpt/tests/css/cssom-view/scrollParent-quirks-mode.html

<html>
<head>
<meta charset="utf-8">
<title>cssom-view - scrollParent in quirks mode</title>
<link rel="help" href="https://drafts.csswg.org/cssom-view/#dom-htmlelement-scrollparent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
test(() => {
  assert_equals(document.scrollingElement, document.body);
  assert_equals(document.body.scrollParent(), null);
}, "scrollParent of body is null when body is document scrolling element");
</script>
</body>
</html>
